+++ api/core/workflow/graph_engine/entities/event.py
@@ -4,8 +4,8 @@ from typing import Any, Optional
 from pydantic import BaseModel, Field
 
 from core.workflow.graph_engine.entities.runtime_route_state import RouteNodeState
+from core.workflow.nodes import NodeType
 from core.workflow.nodes.base import BaseNodeData
-from enums import NodeType
 
 
 class GraphEngineEvent(BaseModel):
