+++ api/core/workflow/nodes/answer/__init__.py
@@ -1,3 +1,4 @@
+from .answer_node import AnswerNode
 from .entities import AnswerStreamGenerateRoute
 
-__all__ = ["AnswerStreamGenerateRoute"]
+__all__ = ["AnswerStreamGenerateRoute", "AnswerNode"]
