+++ api/core/workflow/nodes/answer/answer_node.py
@@ -11,8 +11,8 @@ from core.workflow.nodes.answer.entities import (
     VarGenerateRouteChunk,
 )
 from core.workflow.nodes.base import BaseNode
+from core.workflow.nodes.enums import NodeType
 from core.workflow.utils.variable_template_parser import VariableTemplateParser
-from enums import NodeType
 from models.workflow import WorkflowNodeExecutionStatus
