+++ api/core/workflow/nodes/end/__init__.py
@@ -1,3 +1,4 @@
+from .end_node import EndNode
 from .entities import EndStreamParam
 
-__all__ = ["EndStreamParam"]
+__all__ = ["EndStreamParam", "EndNode"]
