+++ api/core/workflow/nodes/end/end_node.py
@@ -4,7 +4,7 @@ from typing import Any
 from core.workflow.entities.node_entities import NodeRunResult
 from core.workflow.nodes.base import BaseNode
 from core.workflow.nodes.end.entities import EndNodeData
-from enums import NodeType
+from core.workflow.nodes.enums import NodeType
 from models.workflow import WorkflowNodeExecutionStatus
