+++ api/core/workflow/nodes/http_request/node.py
@@ -10,9 +10,9 @@ from core.tools.tool_file_manager import ToolFileManager
 from core.workflow.entities.node_entities import NodeRunResult
 from core.workflow.entities.variable_entities import VariableSelector
 from core.workflow.nodes.base import BaseNode
+from core.workflow.nodes.enums import NodeType
 from core.workflow.nodes.http_request.executor import Executor
 from core.workflow.utils import variable_template_parser
-from enums import NodeType
 from models.workflow import WorkflowNodeExecutionStatus
 
 from .entities import (
