+++ api/core/workflow/nodes/if_else/__init__.py
@@ -1,0 +1,3 @@
+from .if_else_node import IfElseNode
+
+__all__ = ["IfElseNode"]
