+++ api/core/workflow/nodes/if_else/if_else_node.py
@@ -6,10 +6,10 @@ from typing_extensions import deprecated
 from core.workflow.entities.node_entities import NodeRunResult
 from core.workflow.entities.variable_pool import VariablePool
 from core.workflow.nodes.base import BaseNode
+from core.workflow.nodes.enums import NodeType
 from core.workflow.nodes.if_else.entities import IfElseNodeData
 from core.workflow.utils.condition.entities import Condition
 from core.workflow.utils.condition.processor import ConditionProcessor
-from enums import NodeType
 from models.workflow import WorkflowNodeExecutionStatus
