+++ api/core/workflow/nodes/knowledge_retrieval/__init__.py
@@ -1,0 +1,3 @@
+from .knowledge_retrieval_node import KnowledgeRetrievalNode
+
+__all__ = ["KnowledgeRetrievalNode"]
