+++ api/core/workflow/nodes/parameter_extractor/__init__.py
@@ -1,0 +1,3 @@
+from .parameter_extractor_node import ParameterExtractorNode
+
+__all__ = ["ParameterExtractorNode"]
