+++ api/core/workflow/nodes/parameter_extractor/parameter_extractor_node.py
@@ -25,9 +25,9 @@ from core.prompt.simple_prompt_transform import ModelMode
 from core.prompt.utils.prompt_message_util import PromptMessageUtil
 from core.workflow.entities.node_entities import NodeRunMetadataKey, NodeRunResult
 from core.workflow.entities.variable_pool import VariablePool
+from core.workflow.nodes.enums import NodeType
 from core.workflow.nodes.llm import LLMNode, ModelConfig
 from core.workflow.utils import variable_template_parser
-from enums import NodeType
 from extensions.ext_database import db
 from models.workflow import WorkflowNodeExecutionStatus
