+++ api/core/workflow/nodes/question_classifier/__init__.py
@@ -1,3 +1,4 @@
 from .entities import QuestionClassifierNodeData
+from .question_classifier_node import QuestionClassifierNode
 
-__all__ = ["QuestionClassifierNodeData"]
+__all__ = ["QuestionClassifierNodeData", "QuestionClassifierNode"]
