+++ api/core/workflow/nodes/start/__init__.py
@@ -1,0 +1,3 @@
+from .start_node import StartNode
+
+__all__ = ["StartNode"]
