+++ api/core/workflow/nodes/variable_aggregator/__init__.py
@@ -1,0 +1,3 @@
+from .variable_aggregator_node import VariableAggregatorNode
+
+__all__ = ["VariableAggregatorNode"]
