+++ api/core/workflow/nodes/variable_aggregator/variable_aggregator_node.py
@@ -3,8 +3,8 @@ from typing import Any
 
 from core.workflow.entities.node_entities import NodeRunResult
 from core.workflow.nodes.base import BaseNode
+from core.workflow.nodes.enums import NodeType
 from core.workflow.nodes.variable_aggregator.entities import VariableAssignerNodeData
-from enums import NodeType
 from models.workflow import WorkflowNodeExecutionStatus
