+++ api/events/event_handlers/delete_tool_parameters_cache_when_sync_draft_workflow.py
@@ -1,7 +1,7 @@
 from core.tools.tool_manager import ToolManager
 from core.tools.utils.configuration import ToolParameterConfigurationManager
+from core.workflow.nodes import NodeType
 from core.workflow.nodes.tool.entities import ToolEntity
-from enums import NodeType
 from events.app_event import app_draft_workflow_was_synced
