+++ api/events/event_handlers/update_app_dataset_join_when_app_published_workflow_updated.py
@@ -1,7 +1,7 @@
 from typing import cast
 
+from core.workflow.nodes import NodeType
 from core.workflow.nodes.knowledge_retrieval.entities import KnowledgeRetrievalNodeData
-from enums import NodeType
 from events.app_event import app_published_workflow_was_updated
 from extensions.ext_database import db
 from models.dataset import AppDatasetJoin
