+++ api/tests/unit_tests/core/workflow/nodes/answer/test_answer_stream_processor.py
@@ -13,8 +13,8 @@ from core.workflow.graph_engine.entities.event import (
 from core.workflow.graph_engine.entities.graph import Graph
 from core.workflow.graph_engine.entities.runtime_route_state import RouteNodeState
 from core.workflow.nodes.answer.answer_stream_processor import AnswerStreamProcessor
+from core.workflow.nodes.enums import NodeType
 from core.workflow.nodes.start.entities import StartNodeData
-from enums import NodeType
 
 
 def _recursive_process(graph: Graph, next_node_id: str) -> Generator[GraphEngineEvent, None, None]:
